feat: modularize feature of tagging questioner and creating rank shortcut
